[Genomics]: Only move files if they don't already exist.

+++ genomics/sparkcaller/src/main/java/com/github/sparkcaller/Utils.java
@@ -72,7 +72,10 @@ public class Utils {
 
     public static File moveToDir(File fileToCopy, String targetPath) throws IOException {
         File outputFile = new File(targetPath, fileToCopy.getName());
-        FileUtils.moveFile(fileToCopy, outputFile);
+
+        if (!outputFile.exists()) {
+            FileUtils.moveFile(fileToCopy, outputFile);
+        }
 
         return outputFile;
     }
